Move session settings endpoint to /session/settings Add documentation for /session/settings
